+++ crest/Assets/Crest/Crest/Scripts/Underwater/UnderwaterRenderer.cs
@@ -62,6 +62,7 @@ namespace Crest
             public bool _viewOceanMask = false;
             public bool _disableOceanMask = false;
             public bool _disableHeightAboveWaterOptimization = false;
+            public bool _disableArtifactCorrection = false;
         }
 
         Camera _camera;