+++ docs/about/history.rst
@@ -47,6 +47,7 @@ Fixed
    -  Fix FFT spectrum not being editable when time is paused.
    -  Fix *ShapeFFT* component producing inverted looking waves when enabled in editor play mode.
    -  Fix SSS colour missing or popping in the distance.
+   -  Fix underwater artefacts (bright specks).
 
    .. only:: birp